fix:separate file to fix coverage issue
