+++ .github/workflows/lint.yml
@@ -43,7 +43,6 @@ jobs:
       - name: Run Linting
         shell: bash
         run: |
-          rm -f .eslintcache
           if ! pnpm run lint; then
               # print a nice error message on lint failure
               ERROR_MESSAGE='Running `pnpm run lint` failed.'
